EntityManagerFactoryTest split out assets
