+++ tests/Assets/AnotherListenerStub.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace LaravelDoctrineTest\ORM\Assets;
+
+class AnotherListenerStub
+{
+}
