+++ tests/Assets/Decorator.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace LaravelDoctrineTest\ORM\Assets;
+
+use Doctrine\ORM\Decorator\EntityManagerDecorator;
+
+class Decorator extends EntityManagerDecorator
+{
+}
