+++ tests/Assets/FakeConnection.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace LaravelDoctrineTest\ORM\Assets;
+
+use Doctrine\DBAL\Connection;
+
+class FakeConnection extends Connection
+{
+}
