+++ tests/Assets/FakeEventManager.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace LaravelDoctrineTest\ORM\Assets;
+
+class FakeEventManager extends \Doctrine\Common\EventManager
+{
+}
