+++ tests/Assets/SubscriberStub.php
@@ -1,0 +1,19 @@
+<?php
+
+namespace LaravelDoctrineTest\ORM\Assets;
+
+use Doctrine\Common\EventSubscriber;
+
+class SubscriberStub implements EventSubscriber
+{
+    /**
+     * Returns an array of events this subscriber wants to listen to.
+     * @return array
+     */
+    public function getSubscribedEvents()
+    {
+        return [
+            'onFlush'
+        ];
+    }
+}
